requirements.txt
numpy>=1.21.0
pandas>=1.3.0
python-dateutil>=2.8.2
numpy-financial>=1.0.0